fix: move utils folder into src
