+++ src/services/authServices.ts
@@ -1,4 +1,4 @@
-import prisma from '../../utils/client';
+import prisma from '../utils/client';
 import bcrypt from 'bcrypt';
 
 interface UserData {
